refactored derivative files
