+++ csdl_alpha/src/operations/derivative/derivative.py
@@ -1,4 +1,6 @@
 from csdl_alpha.src.operations.operation_subclasses import ElementwiseOperation, ComposedOperation
+from csdl_alpha.src.operations.derivative.reverse import vjp
+from csdl_alpha.src.operations.derivative.bookkeeping import listify_and_verify_variables
 from csdl_alpha.src.graph.operation import Operation, set_properties 
 from csdl_alpha.src.graph.variable import Variable
 from csdl_alpha.src.graph.graph import Graph
@@ -6,210 +8,10 @@ from csdl_alpha.utils.inputs import variablize, validate_and_variablize, get_typ
 import csdl_alpha.utils.testing_utils as csdl_tests
 from csdl_alpha.utils.typing import VariableLike
 
+
 from typing import Union
 
-class Reverse():
 
-    def __init__(self,x:Variable,y:Variable):
-        super().__init__(x,y)
-        self.name = 'reverse'
-
-    def compute_inline(self, x, y):
-        return x + y
-
-    def compute_inline(self, x, y):
-        return x + y
-
-class VarTangents():
-    def __init__(self):
-        """Store cotangents/tangents for variables in the graph.
-        """
-        self.tangent_dictionary:dict[Variable: Variable] = {}
-
-    def accumulate(self, variable:Variable, tangent:Variable)->None:
-        """Accumulate a tangent for a variable.
-
-        Args:
-            variable (Variable): The variable to accumulate the tangent for.
-            tangent (Variable): The tangent to accumulate.
-        """
-        if variable.shape != tangent.shape:
-            raise ValueError(f"Variable shape {variable.shape} and (co)tangent shape {tangent.shape} do not match.")
-        if variable in self.tangent_dictionary:
-            if self.tangent_dictionary[variable] is None:
-                self.tangent_dictionary[variable] = tangent
-            elif isinstance(tangent, Variable):
-                self.tangent_dictionary[variable] = self.tangent_dictionary[variable] + tangent
-            else:
-                raise TypeError(f"Expected tangent to be a Variable, but got {type(tangent)}.")
-        else:
-            raise KeyError(f"Variable {variable} not found in tangent dictionary.")
-
-        self.tangent_dictionary[variable].add_name(f'tangent_{variable.name}')
-
-    def initialize(self, variable:Variable):
-        """Initialize the tangent for a variable.
-
-        Args:
-            variable (Variable): The variable to initialize the tangent for.
-        """
-        if variable not in self.tangent_dictionary:
-            import numpy as np
-            self.tangent_dictionary[variable] = None
-            # self.tangent_dictionary[variable].add_name(f'tangent_{variable.name}')
-
-    def check(self, variable:Variable):
-        """Check if a variable has a tangent.
-
-        Args:
-            variable (Variable): The variable to check.
-
-        Returns:
-            bool: True if the variable has a tangent.
-        """
-        return variable in self.tangent_dictionary
-
-    def __getitem__(self, variable:Variable):
-        """Get the tangent for a variable.
-
-        Args:
-            variable (Variable): The variable to get the tangent for.
-
-        Returns:
-            Variable: The tangent for the variable.
-        """
-        return self.tangent_dictionary[variable]
-
-
-def build_derivative_node_order(
-        graph:Graph,
-        ofs: list[Variable],
-        wrts: list[Variable],
-        reverse: bool = True,
-    )->list[Variable]:
-    """
-    Gets the subgraph of the graph that is relevant for the derivative of the output variables with respect to the input variables.
-    """
-    
-    import rustworkx as rx
-    
-    # Find the subgraph of ofs, wrts
-    intersecting_nodes = graph._get_intersection(
-        wrts,
-        ofs,
-        check_sources=False,
-        check_targets=False,
-        add_hanging_input_variables=False,
-        add_hanging_output_variables=False,
-    )
-    for of_var in ofs:
-        intersecting_nodes.add(graph.node_table[of_var])
-    for wrt_var in wrts:
-        intersecting_nodes.add(graph.node_table[wrt_var])
-    # descendants = set()
-    # wrt_var_indices = set()
-    # for wrt_var in wrts:
-    #     descendants.update(rx.descendants(graph.rxgraph, graph.node_table[wrt_var]))
-    #     wrt_var_indices.add(graph.node_table[wrt_var])
-    # relevant_nodes = rx.ancestors(graph.rxgraph, graph.node_table[of_var]).intersection(descendants)
-    # relevant_nodes.add(graph.node_table[of_var])
-    # relevant_nodes.update(wrt_var_indices)
-
-    # Compute the order of nodes to process
-    node_order = []
-    for node in reversed(rx.topological_sort(graph.rxgraph)):
-        node_index = node
-        if node_index in intersecting_nodes:
-            node_order.append(graph.rxgraph[node])
-    return node_order
-
-def listify_and_verify_variables(variables:Union[Variable, list[Variable]])->list[Variable]:
-    if isinstance(variables, (list, tuple)):
-        variables = [validate_and_variablize(var) for var in variables]
-    else:
-        variables = [validate_and_variablize(variables)]
-    return variables
-
-def vjp(seeds:list[tuple[Variable, Variable]],
-        wrts:Union[Variable, list[Variable]],
-        graph:Graph,
-    )->dict[Variable]:
-    """ Computes the vector-Jacobian product of the seeds with respect to the wrts in the graph.
-
-    Parameters
-    ----------
-    seeds : list[tuple[Variable, Variable]]
-        A list of variable and seed pairs
-    wrts : Union[Variable, list[Variable]]
-        A list of variables to propagate derivatives through
-    graph : Graph
-        The graph in which to compute the derivatives
-
-    Returns
-    -------
-    dict[Variable]
-        The accumulated cotangents for the wrt variables given the output seeds
-
-    Raises
-    ------
-    ValueError
-        Seeds must match the shape of the associated variable
-    """
-
-    # Preprocess inputs
-    of_vars = []
-    for of_var, of_seeds in seeds:
-        of_vars.append(validate_and_variablize(of_var))
-
-        # Seeds must match shape of the variable
-        if of_seeds.shape != of_var.shape:
-            raise ValueError(f"Seed shape {of_seeds.shape} and variable shape {of_var.shape} do not match.")
-    wrt_vars = listify_and_verify_variables(wrts)
-
-    # Extract subgraph of relevant nodes in the graph
-    node_order = build_derivative_node_order(graph, of_vars, wrt_vars)
-
-    # initialize seeds and final wrt cotangents
-    import numpy as np
-    import csdl_alpha as csdl
-
-    cotangents = VarTangents()
-    for i, of_var in enumerate(of_vars):
-        cotangents.initialize(of_var)
-        cotangents.accumulate(of_var, seeds[i][1])
-
-    # perform the vector-jacobian here in terms of CSDL operations by going through the node order
-    for node in node_order:
-        if isinstance(node, Variable):
-            cotangents.initialize(node)
-
-    for node in node_order:
-        # rec = csdl.get_current_recorder()
-        # rec.visualize_graph(filename = graph.name)
-        if isinstance(node, Operation):
-            # Moved to composed operation class for now
-            # for output in node.outputs:
-            #     if cotangents[output] is None:
-            #         cotangents.accumulate(output, csdl.Variable(value = np.zeros(output.shape)))
-            
-            # for output in node.outputs:
-            #     if cotangents[output] is None:
-            #         print(f"cotangents[{output}] is None")
-            try:
-                node.evaluate_vjp(cotangents, *node.inputs, *node.outputs)
-            except Exception as e:
-                raise ValueError(f"Error with VJP in operation {node.name}: {e}")
-
-
-    wrt_cotangents:dict[Variable:Variable] = {}
-    for wrt_var in wrt_vars:
-        wrt_cotangent = cotangents[wrt_var]
-        if wrt_cotangent is None:
-            wrt_cotangents[wrt_var] = None
-            continue
-            # wrt_cotangent = csdl.Variable(name = f'seeds_{wrt_var.name}', value = np.zeros(wrt_var.shape))
-        wrt_cotangents[wrt_var] = wrt_cotangent
-    return wrt_cotangents
 
 def reverse(
         of: Variable,
